+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_04_01_104627) do
+ActiveRecord::Schema[8.1].define(version: 2026_04_07_094142) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -326,6 +326,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_04_01_104627) do
   end
 
   create_table "survey_sends", force: :cascade do |t|
+    t.datetime "completed_at"
     t.datetime "created_at", null: false
     t.datetime "sent_at", null: false
     t.bigint "survey_id", null: false