+++ test/jobs/send_survey_reminder_job_test.rb
@@ -1,6 +1,6 @@
 require "test_helper"
 
-class SendSurveyJobTest < ActiveSupport::TestCase
+class SendSurveyReminderJobTest < ActiveSupport::TestCase
   include ActiveJob::TestHelper
   include Rails.application.routes.url_helpers
 
@@ -9,11 +9,11 @@ class SendSurveyJobTest < ActiveSupport::TestCase
     survey = create(:survey)
 
     assert_enqueued_jobs 1, only: SendCustomMessageJob do
-      SendSurveyJob.new.perform(user, survey)
+      SendSurveyReminderJob.new.perform(user, survey)
     end
 
     assert_equal 1, Message.count
-    assert_match("survey", Message.last.body)
+    assert_match("/surveys/", Message.last.body)
     assert SurveySend.exists?(user: user, survey: survey)
   end
 
@@ -22,11 +22,11 @@ class SendSurveyJobTest < ActiveSupport::TestCase
     survey = create(:survey)
 
     assert_enqueued_jobs 1, only: SendCustomMessageJob do
-      SendSurveyJob.new.perform(user, survey)
+      SendSurveyReminderJob.new.perform(user, survey)
     end
 
     assert_equal 1, Message.count
-    assert_match("arolwg", Message.last.body)
+    assert_match("Cymerwch funud i", Message.last.body)
     assert SurveySend.exists?(user: user, survey: survey)
   end
 
@@ -36,21 +36,9 @@ class SendSurveyJobTest < ActiveSupport::TestCase
 
     User.any_instance.stubs(:generate_token_for).with(:survey_token).returns("abc123")
 
-    SendSurveyJob.new.perform(user, survey)
+    SendSurveyReminderJob.new.perform(user, survey)
 
     assert_match edit_survey_url(survey, token: "abc123"), Message.last.body
-  end
-
-  test "#perform does not send if survey already sent to user" do
-    user = create(:user)
-    survey = create(:survey)
-    create(:survey_send, user: user, survey: survey)
-
-    assert_no_enqueued_jobs only: SendCustomMessageJob do
-      SendSurveyJob.new.perform(user, survey)
-    end
-
-    assert_equal 0, Message.count
   end
 
   test "#perform does not create SurveySend if message fails to save" do
@@ -60,7 +48,7 @@ class SendSurveyJobTest < ActiveSupport::TestCase
     Message.any_instance.stubs(:save).returns(false)
 
     assert_no_enqueued_jobs only: SendCustomMessageJob do
-      SendSurveyJob.new.perform(user, survey)
+      SendSurveyReminderJob.new.perform(user, survey)
     end
 
     assert_equal 0, SurveySend.count