+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## [2026-04-07]
+- Put a link to pre-programme survey on the thank you page
+- Send reminders to users to fill it in if they haven't a couple of days later
+
 ## [2026-02-04]
 - Close sign ups before pilot in Wales